+++ library/SolverUtils/DriverStandard.cpp
@@ -34,6 +34,7 @@
 
 #include <iomanip>
 
+#include <LibUtilities/BasicUtils/Timer.h>
 #include <SolverUtils/DriverStandard.h>
 
 using namespace std;
@@ -75,23 +76,23 @@ void DriverStandard::v_InitObject(ostream &out)
 void DriverStandard::v_Execute(ostream &out)
 
 {
-    time_t starttime, endtime;
+    Nektar::LibUtilities::Timer timer;
     NekDouble CPUtime;
 
     m_equ[0]->PrintSummary(out);
 
-    time(&starttime);
+    timer.Start();
 
     m_equ[0]->DoInitialise();
     m_equ[0]->DoSolve();
 
-    time(&endtime);
+    timer.Stop();
 
     m_equ[0]->Output();
 
     if (m_comm->GetRank() == 0)
     {
-        CPUtime = difftime(endtime, starttime);
+        CPUtime = timer.Elapsed().count();
         cout << "-------------------------------------------" << endl;
         cout << "Total Computation Time = " << CPUtime << "s" << endl;
         cout << "-------------------------------------------" << endl;